Add godot.enums and godot.global_functions modules
